Let doit take care of keeping dependencies in sync
Since doit itself is managed as a dependency, it cannot take care of
initial setup. Syncing dependencies after a git pull, for instance,
should just work when doit is executed. It may be a good idea to
import modules not at the beginning of dodo.py, but in the functions
depending on those modules, instead. This way, sync_dependencies gets
a chance to pull everything in before import can raise an exception.