Build schema documentation on demand
The operation is not idem potent, apparently. The order of elements in
the generated markdown files is not stable from one call to another.
This is not a big deal as long as all information is included in the
end. It does make this task unsuitable for unconditional execution,
e.g. as a pre-commit hook, though. Just regenerate documentation when
appropriate by executing doit docs, for now.