Make has_record single-valued
Do not model a collection of moving image metadata records, but only a
single instance of such a record. This ensures that every instance
validating against the source schema will also validate against one of
the ePIC kernel inforamiotn profiles for work, manifestation or item.
In particular, moving image records produced with the derived python
bindings should be ready for the ePIC API without unpacking them from
a list, now.