Make provisions for temporary local identifiers
Introduce identifiers of local scope, so data providers can use the
AVefi schema as an export format, even if they do not communicate
directly with the ePIC API.

Explicitly mark all classes that are to be implemented as InfoTypes in
the type registry. Local identifiers are not part of this subset since
they should never make it into the PID system.